move more indexing tests
